Added my test javas
This stuff is just using the jsoup to get that working. My Event class
is almost identical to Amanda's. I have all the events being saved into
an object and then I add the Event object to an ArrayList. Almsot is
working OK. I can't get the title right now but I'll fix that soon.